Agregue los metodos de pago y sus API

+++ Ecommerce/src/envio/MetodoDeEnvio.java
@@ -5,6 +5,6 @@ import ecommerce.Pedido;
 public interface MetodoDeEnvio {
 
 	
-	 double costoDeEnvio(Pedido pedido);
+	 float costoDeEnvio(Pedido pedido);
 	
 }
